backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-dotenv==1.0.0
psycopg[binary]==3.2.3
sqlalchemy==2.0.36
pydantic==2.10.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
googlemaps==4.10.0
sendgrid==6.11.0
email-validator==2.1.0
python-multipart==0.0.20
